rename DataSourceAPIPandas to DataFrame and fix specs
