+++ a2ml/api/auger/impl/mparts/predict.py
@@ -8,8 +8,7 @@ from ..cloud.cluster import AugerClusterApi
 from ..cloud.pipeline import AugerPipelineApi
 from ..exceptions import AugerException
 from a2ml.api.utils import fsclient
-#from a2ml.api.utils.dataframe import DataFrame
-from a2ml.api.utils.data_source_api_pandas import DataSourceAPIPandas
+from a2ml.api.utils.dataframe import DataFrame
 from a2ml.api.review_model.model_helper import ModelHelper
 
 class ModelPredict():
@@ -32,33 +31,17 @@ class ModelPredict():
         return predicted
 
     def _predict_on_cloud(self, filename, model_id, threshold, data, columns, output):
-        ds = DataSourceAPIPandas.create_dataframe(filename, data, columns)
+        ds = DataFrame.create_dataframe(filename, data, columns)
 
         pipeline_api = AugerPipelineApi(self.ctx, None, model_id)
         predictions = pipeline_api.predict(ds.get_records(), ds.columns, threshold)
 
-        ds_result = DataSourceAPIPandas.create_dataframe(None, records=predictions['data'], features=predictions['columns'])
+        ds_result = DataFrame.create_dataframe(None, records=predictions['data'], features=predictions['columns'])
         ds_result.options['data_path'] = filename
         return ModelHelper.save_prediction_result(ds_result, 
             prediction_id = None, support_review_model = False, 
             json_result=False, count_in_result=False, prediction_date=None, 
             model_path=None, model_id=model_id, output=output)
-
-        # target = self.ctx.config.get('target', None)
-        # records, features = DataFrame.load_records(filename, target, features=columns, data=data)
-
-        # pipeline_api = AugerPipelineApi(self.ctx, None, model_id)
-        # predictions = pipeline_api.predict(records, features, threshold)
-
-        # if filename:
-        #     predicted = os.path.splitext(filename)[0] + "_predicted.csv"
-        #     DataFrame.save(predicted, predictions)
-        # elif columns:
-        #     predicted = predictions.get('data', [])
-        # else:
-        #     predicted = DataFrame.convert_records_to_dict(predictions)
-
-        # return predicted
 
     def _predict_locally(self, filename_arg, model_id, threshold, data, columns, output):
         model_deploy = ModelDeploy(self.ctx, None)
@@ -73,7 +56,7 @@ class ModelPredict():
 
         filename = filename_arg
         if not filename:
-            ds = DataSourceAPIPandas.create_dataframe(filename, data, columns)            
+            ds = DataFrame.create_dataframe(filename, data, columns)            
             filename = os.path.join(self.ctx.config.get_path(), '.augerml', 'predict_data.csv')
             ds.saveToCsvFile(filename, compression=None)
 
@@ -86,16 +69,19 @@ class ModelPredict():
             if not model_existed:
                 shutil.rmtree(model_path, ignore_errors=True)
 
-        ds_result = DataSourceAPIPandas.create_dataframe(predicted)
         if not filename_arg:
+            ds_result = DataFrame.create_dataframe(predicted)
+
             ds_result.options['data_path'] = None
             ds_result.loaded_columns = columns
 
-        return ModelHelper.save_prediction_result(ds_result, 
-            prediction_id = None, support_review_model = False, 
-            json_result=False, count_in_result=False, prediction_date=None, 
-            model_path=None, model_id=model_id, output=output)
+            return ModelHelper.save_prediction_result(ds_result, 
+                prediction_id = None, support_review_model = False, 
+                json_result=False, count_in_result=False, prediction_date=None, 
+                model_path=None, model_id=model_id, output=output)
 
+        return predicted
+            
     def _extract_model(self, model_name):
         model_path = os.path.splitext(model_name)[0]
         model_existed = os.path.exists(model_path)
@@ -129,6 +115,7 @@ class ModelPredict():
                 'Running model in deeplearninc/'
                 'auger-ml-worker:%s' % docker_tag)
             result_file = subprocess.check_output(command, stderr=subprocess.STDOUT, shell=True)
+            print(result_file)
             result_file = result_file.decode("utf-8").strip()
             result_file = os.path.basename(result_file)
             # getattr(subprocess,
